Combine calculators into one
