+++ apps/web/components/connected/logreglan/FineAndSpeedMeasurementCalculator/FineAndSpeedMeasurementCalculator.css.ts
