+++ apps/web/components/connected/logreglan/FineAndSpeedMeasurementCalculator/index.ts
@@ -1,0 +1,1 @@
+export { FineAndSpeedMeasurementCalculator } from './FineAndSpeedMeasurementCalculator'
