+++ apps/web/components/connected/logreglan/FineAndSpeedMeasurementCalculator/utils.ts
