--- apps/web/components/connected/logreglan/FineCalculator/index.ts
@@ -1,1 +1,0 @@
-export * from './FineCalculator'
