--- apps/web/components/connected/logreglan/SpeedMeasurementCalculator/index.ts
@@ -1,1 +1,0 @@
-export * from './SpeedMeasurementCalculator'
