--- apps/web/components/connected/logreglan/SpeedMeasurementCalculator/translation.strings.ts
@@ -1,69 +1,0 @@
-import { defineMessages } from 'react-intl'
-
-export const m = defineMessages({
-  measuredSpeedInputLabel: {
-    id: 'web.logreglan.speedMeasurementCalculator:measuredSpeedInputLabel',
-    defaultMessage: 'Mældur hraði',
-    description: 'Mældur hraði',
-  },
-  vikmorkInputLabel: {
-    id: 'web.logreglan.speedMeasurementCalculator:vikmorkInputLabel',
-    defaultMessage: 'Vikmörk',
-    description: 'Vikmörk',
-  },
-  nidurstadaInputLabel: {
-    id: 'web.logreglan.speedMeasurementCalculator:nidurstadaInputLabel',
-    defaultMessage: 'Niðurstaða',
-    description: 'Niðurstaða',
-  },
-  speedLimitSelectLabel: {
-    id: 'web.logreglan.speedMeasurementCalculator:speedLimitSelectLabel',
-    defaultMessage: 'Hámarkshraði',
-    description: 'Hámarkshraði',
-  },
-  yes: {
-    id: 'web.logreglan.speedMeasurementCalculator:yes',
-    defaultMessage: 'Já',
-    description: 'Já',
-  },
-  no: {
-    id: 'web.logreglan.speedMeasurementCalculator:no',
-    defaultMessage: 'Nei',
-    description: 'Nei',
-  },
-  over3500kgOrWithTrailerLabel: {
-    id: 'web.logreglan.speedMeasurementCalculator:over3500kgOrWithTrailerLabel',
-    defaultMessage: 'Bifreiðar þyngri en 3500kg eða með eftirvagn?',
-    description: 'Bifreiðar þyngri en 3500kg eða með eftirvagn?',
-  },
-  finePrefix: {
-    id: 'web.logreglan.speedMeasurementCalculator:finePrefix',
-    defaultMessage: 'Sekt: ',
-    description: 'Sekt: ',
-  },
-  pointsPrefix: {
-    id: 'web.logreglan.speedMeasurementCalculator:pointsPrefix',
-    defaultMessage: 'Punktar: ',
-    description: 'Punktar: ',
-  },
-  twentyPercentLoadPrefix: {
-    id: 'web.logreglan.speedMeasurementCalculator:twentyPercentLoadPrefix',
-    defaultMessage: '20% álag: ',
-    description: '20% álag: ',
-  },
-  monthsOfDrivingLicenseLossPrefix: {
-    id: 'web.logreglan.speedMeasurementCalculator:monthsOfDrivingLicenseLossPrefix',
-    defaultMessage: 'Svipting ökuréttinda í: ',
-    description: 'Svipting ökuréttinda í: ',
-  },
-  months: {
-    id: 'web.logreglan.speedMeasurementCalculator:months',
-    defaultMessage: 'mánuði',
-    description: 'mánuði',
-  },
-  akaeraText: {
-    id: 'web.logreglan.speedMeasurementCalculator:akaeraText',
-    defaultMessage: 'Ákært verður vegna brotsins',
-    description: 'Ákært verður vegna brotsins',
-  },
-})
